+++ CHANGELOG.md
@@ -5,6 +5,15 @@
 **Please go to the add-on configuration page and click "Restore Defaults". Sorry for the inconvenience!**
 
 
+### Version **1.0.5** 2023-06-15
+
+Added
+
+* System prompt.
+
+Changed
+
+* Now model is drop down menu.
 
 ### Version **1.0.4** 2023-06-12